added fitnessScore function

+++ src/Travel.h
@@ -2,6 +2,6 @@
 #define Travel_H
 #include "CityCoordinate.h"
 
-double fitnessScore(City city1, City city2);
+float fitnessScore(City city1, City city2);
 
 #endif // Travel_H
